Create README.md
In the following project, a Lego Mindstorm EV3 robot kit was used to build and program a Surface Cleaning Robot. The robot can be used to clean the surfaces of tables, using a revolving cylindrical surface cleaning scrubber. The robot is integrated with touch sensors at the front that allows edge detection which prevents the robot from falling off the edge. It also comprises of a color sensor and periodic cleaner spray that detects the amount of spray present in the canister. The additional functionality of obstacle avoidance is also integrated into the robot using an ultrasonic distance sensor. Algorithms to clean circular and Rectangular shaped tables have been included in the programming section. For tables with irregular shapes, the robot goes along the edge to determine the shape of the table between engaging in the cleaning process. The program was written on RobotC programming language. Three more members were involved in the development of the project.